refactor(test): migrate custom symbol matchers to shared utils package
