+++ packages/utils/src/equal-ignoring-symbols.ts
